Do not calculate the length of the IEND element
Due to changes in Daffodil's alignment algorithm, we get stuck in a
deadlock when trying to calculate the length of the IEND element when
unparser. However, the length of IEND is always zero, so we can just
hardcode a zero for the output value calc rather than trying to figure
it out, which avoids the deadlock.